Moved newly created classes to Core Utilities
