+++ CoreUtilities/test/unit/src/au/gov/asd/tac/constellation/utilities/gui/SelectOptionsExtensionNGTest.java
@@ -13,7 +13,7 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
  */
-package au.gov.asd.tac.constellation.plugins.parameters;
+package au.gov.asd.tac.constellation.utilities.gui;
 
 import java.util.ArrayList;
 import java.util.List;
